fix(solver): preserve wrapper alias display
